+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes relevant to players in this project will be documented in th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 ## [Unreleased]
+### Added
+- Protection History (on daily bonus page during protection) is now formatted to be imported as an Excel sim log.
 
 ## [1.37.1] - 2023-10-15
 ### Changed